agenda no padrão finalizada

+++ app/Http/Controllers/ContactsController.php
@@ -2,11 +2,16 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Contacts;
+use App\Events\NewContact;
 use App\Http\Requests\ContactsRequest;
+use App\Mail\NewContactMail;
 use App\Services\AddressesService;
 use App\Services\ContactsService;
 use App\Services\EmailService;
+use App\Services\Params\CreateAddressServiceParams;
+use App\Services\Params\CreateContactServiceParams;
+use App\Services\Params\CreatePhoneServiceParams;
+use App\Services\Params\UpdateContactServiceParams;
 use App\Services\PhonesService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -32,46 +37,81 @@ class ContactsController extends Controller
         $this->middleware('auth');
     }
 
-
-
-    public function show($id, $returnView = false)
+    public function show($id)
     {
-        $contact = $this->contactsService->find($id);
+        $response = $this->contactsService->find($id);
 
-        if (!$contact) {
-            return ResponseService::returnApi(false, null, "Contato não encontrado");
+        if (!$response->success) {
+            return back()->with('warning', $response->message);
         }
 
-        $phones = $this->phonesService->searchPhones($contact->contact_id);
+        $contact = $response->data;
+        $phonesResponse = $this->phonesService->searchPhones($contact->contact_id);
+        $addressesResponse = $this->addressesService->searchAddresses($contact->contact_id);
 
-        $addresses = $this->addressesService->searchAddresses($contact->contact_id);
-
-        $contact->phones = $phones;
-        $contact->addresses = $addresses;
-
-        if (!$returnView) {
-            return view('contact-form', compact('contact'));
+        if (!$phonesResponse->success) {
+            return back()->with('warning', $phonesResponse->message);
+        }
+        if (!$addressesResponse->success) {
+            return back()->with('warning', $addressesResponse->message);
         }
 
-        return $contact;
+        $contact->phones = $phonesResponse->data;
+        $contact->addresses = $addressesResponse->data;
+
+        return view('contact-modal', compact('contact'));
+    }
+
+    public function edit($id)
+    {
+        $response = $this->contactsService->find($id);
+
+        if (!$response->success) {
+            return back()->with('warning', $response->message);
+        }
+
+        $contact = $response->data;
+        $phonesResponse = $this->phonesService->searchPhones($contact->contact_id);
+        $addressesResponse = $this->addressesService->searchAddresses($contact->contact_id);
+
+        if (!$phonesResponse->success) {
+            return back()->with('warning', $phonesResponse->message);
+        }
+        if (!$addressesResponse->success) {
+            return back()->with('warning', $addressesResponse->message);
+        }
+
+        $contact->phones = $phonesResponse->data;
+        $contact->addresses = $addressesResponse->data;
+
+        return view('contact-form', compact('contact'));
     }
 
     public function index()
     {
-        $id = Auth::id();
-        $contacts = $this->contactsService->searchContacts($id);
 
-        if (!$contacts) {
-            return back()->with('warning', 'Contatos não encontrados');
+        $id = Auth::id();
+        $response = $this->contactsService->searchContacts($id);
+
+        if (!$response->success) {
+            return back()->with('warning', $response->message);
         }
 
+        $contacts = $response->data;
+
         foreach ($contacts as $contact) {
-            $phones = $this->phonesService->searchPhones($contact->contact_id);
+            $phonesResponse = $this->phonesService->searchPhones($contact->contact_id);
+            $addressesResponse = $this->addressesService->searchAddresses($contact->contact_id);
 
-            $addresses = $this->addressesService->searchAddresses($contact->contact_id);
+            if (!$phonesResponse->success) {
+                return back()->with('warning', $phonesResponse->message);
+            }
+            if (!$addressesResponse->success) {
+                return back()->with('warning', $addressesResponse->message);
+            }
 
-            $contact->phones = $phones;
-            $contact->addresses = $addresses;
+            $contact->phones = $phonesResponse->data;
+            $contact->addresses = $addressesResponse->data;
         }
 
         return view('contacts', compact('contacts'));
@@ -80,26 +120,66 @@ class ContactsController extends Controller
     public function store(ContactsRequest $request)
     {
         try {
-            $contact = $this->contactsService->store($request);
+            $id = Auth::id();
+            $contactParams = new CreateContactServiceParams(
+                $request->name,
+                $id,
+                $request->company,
+                $request->role,
+            );
+
+            $response = $this->contactsService->createContact($contactParams);
+
+            if (!$response->success) {
+                return back()->with('warning', $response->message);
+            }
+
+            $contact = $response->data;
 
             $phoneArray = [];
             foreach ($request->phones as $phone) {
-                $newPhone = $this->phonesService->store($phone, $contact->contact_id);
-                array_push($phoneArray, $newPhone);
+                $phoneParams = new CreatePhoneServiceParams(
+                    $phone['type'],
+                    $phone['phone'],
+                    $contact->contact_id,
+                );
+                $response = $this->phonesService->createPhone($phoneParams);
+
+                if (!$response->success) {
+                    return back()->with('warning', $response->message);
+                }
+
+                array_push($phoneArray, $response->data);
             }
 
             $addressArray = [];
             foreach ($request->addresses as $address) {
-                $newAddress = $this->addressesService->store($address, $contact->contact_id);
-                array_push($addressArray, $newAddress);
+                $addressParams = new CreateAddressServiceParams(
+                    $address['zip_code'],
+                    $address['street'],
+                    $address['number'],
+                    $address['neighborhood'],
+                    $address['complement'],
+                    $address['city'],
+                    $address['state'],
+                    $contact->contact_id,
+                );
+                $response = $this->addressesService->createAddress($addressParams);
+
+                if (!$response->success) {
+                    return back()->with('warning', $response->message);
+                }
+                array_push($addressArray, $response->data);
             }
 
             $contact->phones = $phoneArray;
             $contact->addresses = $addressArray;
+
+            $user = Auth::user();
+            event(new NewContact($contact, $user->email));
         } catch (Throwable $e) {
             return back()->with('error', $e);
         }
-
 
         return redirect()->route('contacts.index')->with('success', 'salvo com sucesso!');
     }
@@ -107,20 +187,56 @@ class ContactsController extends Controller
     public function update(ContactsRequest $request, $id)
     {
         try {
-            $contact = $this->contactsService->update($request, $id);
+            $contactParams = new UpdateContactServiceParams(
+                $request->name,
+                $request->company,
+                $request->role
+            );
+            $response = $this->contactsService->updateContact($contactParams, $id);
+
+            if (!$response->success) {
+                return back()->with('warning', $response->message);
+            }
+
+            $contact = $response->data;
 
             $phoneArray = [];
             $this->phonesService->deleteContactPhones($contact->contact_id);
             foreach ($request->phones as $phone) {
-                $newPhone = $this->phonesService->store($phone, $contact->contact_id);
-                array_push($phoneArray, $newPhone);
+                $phoneParams = new CreatePhoneServiceParams(
+                    $phone['type'],
+                    $phone['phone'],
+                    $contact->contact_id,
+                );
+                $response = $this->phonesService->createPhone($phoneParams);
+                if (!$response->success) {
+                    return back()->with('warning', $response->message);
+                }
+
+                array_push($phoneArray, $response->data);
             }
 
             $addressArray = [];
             $this->addressesService->deleteContactAddresses($contact->contact_id);
             foreach ($request->addresses as $address) {
-                $newAddress = $this->addressesService->store($address, $contact->contact_id);
-                array_push($addressArray, $newAddress);
+                $addressParams = new CreateAddressServiceParams(
+                    $address['zip_code'],
+                    $address['street'],
+                    $address['number'],
+                    $address['neighborhood'],
+                    $address['complement'],
+                    $address['city'],
+                    $address['state'],
+                    $contact->contact_id,
+                );
+
+                $response = $this->addressesService->createAddress($addressParams);
+
+                if (!$response->success) {
+                    return back()->with('warning', $response->message);
+                }
+
+                array_push($addressArray, $response->data);
             }
 
             $contact->phones = $phoneArray;
@@ -140,19 +256,23 @@ class ContactsController extends Controller
      * @param $id
      * @return array
      */
-    public function deleteContact($id)
+    public function destroy($id)
     {
-        $contact = $this->contactsService->find($id);
+        $response = $this->contactsService->find($id);
 
-        if (!$contact) {
-            return "Contato não encontrado";
+        if (!$response->success) {
+            return back()->with('warning', $response->message);
         }
 
-        $this->phonesService->deleteContactPhones($id);
+        $phonesResponse = $this->phonesService->deleteContactPhones($id);
 
-        $this->addressesService->deleteContactAddresses($id);
+        $addressesResponse = $this->addressesService->deleteContactAddresses($id);
 
-        $this->contactsService->deleteContact($id);
+        $response = $this->contactsService->deleteContact($id);
+
+        if (!$response->success || !$phonesResponse->success || !$addressesResponse->success) {
+            return back()->with('warning', 'Erro ao deletar contato, tente novamente');
+        }
 
         return redirect()->route('contacts.index')->with('success', 'deletado com sucesso!');
     }
@@ -161,4 +281,29 @@ class ContactsController extends Controller
     {
         return view('contact-form');
     }
+
+    // public function testMail()
+    // {
+    //     $response = $this->contactsService->find(48);
+
+    //     if (!$response->success) {
+    //         return back()->with('warning', $response->message);
+    //     }
+
+    //     $contact = $response->data;
+    //     $phonesResponse = $this->phonesService->searchPhones($contact->contact_id);
+    //     $addressesResponse = $this->addressesService->searchAddresses($contact->contact_id);
+
+    //     if (!$phonesResponse->success) {
+    //         return back()->with('warning', $phonesResponse->message);
+    //     }
+    //     if (!$addressesResponse->success) {
+    //         return back()->with('warning', $addressesResponse->message);
+    //     }
+
+    //     $contact->phones = $phonesResponse->data;
+    //     $contact->addresses = $addressesResponse->data;
+
+    //     return new NewContactMail($contact);
+    // }
 }
